chore: adding modules forums
